temp

+++ backend/src/main/java/com/fribbels/model/BailiInfo.java
@@ -13,7 +13,7 @@ public class BailiInfo {
     private String details = "";
     private int score = 0;
     private int firstSpeedScore = 0;
-    private int speedScore = 0;
+    private double speedScore = 0;
     private int tankScore = 0;
-    private int dpsScore = 0;
+    private double dpsScore = 0;
 }
